less logging if devmode is not enabled
