+++ package.json
@@ -73,7 +73,7 @@
     "marked": "^13.0.0",
     "pako": "^2.1.0",
     "pocketbase": "^0.21.2",
-    "tone": "^14.9.17",
+    "tone": "^15.1.2",
     "tsparticles": "^3.4.0",
     "workbox-build": "^7.1.1",
     "workbox-expiration": "^7.1.0"
